Bug-fixing
Current problem:
Messenger is an asset and its events are plugged with other assets (Orchestrator manager, etc.)
But when the messenger actually emits events, these are not caught by the instances in the scene, but by other instances attached to the Messenger.
Hint: attach the event handlers to the messenger's events at runtime.